+++ pennylane/pulse/__init__.py
@@ -15,6 +15,6 @@
 This module contains classes and functions used in pulse programming.
 """
 
-from .convenience_functions import constant, rect
+from .convenience_functions import constant, rect, pwc, pwc_from_function
 from .parametrized_evolution import ParametrizedEvolution
 from .parametrized_hamiltonian import ParametrizedHamiltonian
