+++ tests/tests_passing_pylint.txt
@@ -5,10 +5,7 @@ tests/templates/test_subroutines/test_basis_rotation.py
 tests/ops/op_math/test_controlled_ops.py
 tests/ops/op_math/test_controlled.py
 tests/ops/op_math/test_dot.py
-tests/ops/pulse/**
 tests/ops/functions/test_evolve.py
-tests/test_operation.py
 tests/ops/qutrit/test_qutrit_observables.py
-tests/pulse/test_parametrized_hamiltonian.py
-tests/pulse/test_parametrized_evolution.py
-tests/pulse/test_convenience_functions.py
+tests/test_operation.py
+tests/pulse/**
